discord_producer/requirements.txt
kafka-python==2.0.2
discord.py==2.2.2
python-dotenv==1.0.0
requests==2.26.0
